feat: measure step duration (#198) (#233)
Deferred from M8 by ADR-0025 decision 2. "This workflow is slow" is a
complaint; "this step is slow" is something an operator can act on, and
history recorded the durations per instance without ever aggregating
them.

flowdeck.steps.duration records one observation per *execution*, so a
step retried three times contributes three. Averaging them into one would
hide the thing worth seeing - that it took three goes. Tagged with the
outcome, because a step that fails fast and a step that succeeds slowly
are different problems and should not share a series.

Seconds, not milliseconds. Prometheus and the OpenTelemetry conventions
both use base units, and a histogram named in the wrong one is a
dashboard nobody can compare against anything else. The engine's logs
stay in milliseconds, where a human reads them.

This is the first instrument FlowDeck emits that is not a counter, so it
brought the histogram rendering #189 declined to invent - correctly, at
the time, because inventing a bucket format before there was anything to
put in it would have been guessing at the shape.

Bucket edges live on EngineMetrics and the endpoint reads them from
there. Two lists would drift, and the drift would be silent: the
histogram would still render, with buckets that did not match what was
measured.

Two rendering rules are easy to get wrong and impossible to notice:
buckets are cumulative - each edge holds everything at or below it - and
+Inf is required and carries the total. Both are asserted, and both were
verified by mutation. Rendering per-bucket counts fails one scenario.

The +Inf rule needed a scenario of its own. With every observation inside
the edges, +Inf and the final cumulative bucket are equal, so a renderer
that confused them passed everything. A scenario now records a minute-long
step - directly, since one cannot honestly take a minute - and asserts it
lands in no bucket and still in the total. That mutation now fails too.

Two existing assertions needed correcting rather than working around. The
lifecycle-counter scenario asserted a total measurement count, which was
right when lifecycle counters were the only instrument and now breaks
whenever an unrelated one is added; it is scoped to that family. And the
observability guide's "not measured" list no longer names step duration,
so the scenario asserting it does not either - asserting it would now
assert the opposite of the truth.

The test capture only listened for long measurements, so it saw nothing
at all when the first histogram arrived. It takes doubles now.

+++ tests/specs/FlowDeck.Specs/Steps/RetryMetricsSteps.cs
@@ -70,7 +70,7 @@ public sealed class RetryMetricsSteps(EngineContext world)
 
         // Two, not three. The third attempt succeeded, and counting it would
         // make a step that recovered look the same as one still failing.
-        Assert.Equal(2, this.Retries().Sum(measurement => measurement.Value));
+        Assert.Equal(2, (long)this.Retries().Sum(measurement => measurement.Value));
 
     [Then("a workflow that never retried reports nothing")]
     public void ThenNoRetriesForTheOtherStep() =>
@@ -81,14 +81,14 @@ public sealed class RetryMetricsSteps(EngineContext world)
     {
         // Which step is the part an operator cannot get anywhere else without
         // reading history per instance.
-        Assert.Equal(1, this.Retries().Where(m => Equals(m.Tags["step.name"], "reserve")).Sum(m => m.Value));
-        Assert.Equal(1, this.Retries().Where(m => Equals(m.Tags["step.name"], "charge")).Sum(m => m.Value));
+        Assert.Equal(1, (long)this.Retries().Where(m => Equals(m.Tags["step.name"], "reserve")).Sum(m => m.Value));
+        Assert.Equal(1, (long)this.Retries().Where(m => Equals(m.Tags["step.name"], "charge")).Sum(m => m.Value));
     }
 
     [Then("two compensating actions are counted, both as undone")]
     public void ThenTwoUndone()
     {
-        Assert.Equal(2, this.Compensations().Sum(m => m.Value));
+        Assert.Equal(2, (long)this.Compensations().Sum(m => m.Value));
         Assert.All(this.Compensations(), m => Assert.Equal("undone", m.Tags["outcome"]));
     }
 
@@ -97,8 +97,8 @@ public sealed class RetryMetricsSteps(EngineContext world)
     {
         // The difference between "one undo failed" and "nine did", which the
         // per-instance CompensationFailed counter cannot express.
-        Assert.Equal(1, this.Compensations().Where(m => Equals(m.Tags["outcome"], "undone")).Sum(m => m.Value));
-        Assert.Equal(1, this.Compensations().Where(m => Equals(m.Tags["outcome"], "failed")).Sum(m => m.Value));
+        Assert.Equal(1, (long)this.Compensations().Where(m => Equals(m.Tags["outcome"], "undone")).Sum(m => m.Value));
+        Assert.Equal(1, (long)this.Compensations().Where(m => Equals(m.Tags["outcome"], "failed")).Sum(m => m.Value));
 
         Assert.Equal(
             "reserve",
